document_intelligence/fast_latex/requirements.txt
uvicorn
pillow
numpy
opencv-python-headless
onnxruntime
paddleocr
fastapi
git+https://github.com/j2whiting/texteller.git
